requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
fake-useragent==1.1.3
firebase-admin==6.1.0
Flask==1.1.2
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.2
msgpack==1.0.5
oauth2client==3.0.0
proto-plus==1.22.2
protobuf==4.22.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2023.3
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
